refactor: move unittest locations
