--- src/config/tests.rs
@@ -1,6 +1,0 @@
-use super::*;
-
-#[test_case]
-fn test_simple_config() {
-    // TODO: rewrite test with new gicv2 and gicv3 config system
-}
