+++ src/zone.rs
@@ -16,9 +16,6 @@ use crate::memory::{MMIOConfig, MMIOHandler, MMIORegion, MemorySet};
 use crate::percpu::{get_cpu_data, this_zone, CpuSet};
 use core::panic;
 
-#[cfg(test)]
-pub mod tests;
-
 pub struct Zone {
     pub name: [u8; CONFIG_NAME_MAXLEN],
     pub id: usize,
@@ -234,4 +231,19 @@ pub struct ZoneInfo {
     zone_id: u32,
     cpus: u64,
     name: [u8; CONFIG_NAME_MAXLEN],
+}
+
+#[test_case]
+fn test_add_and_remove_zone() {
+    let zone_count = 50;
+    let zone_count_before = ZONE_LIST.read().len();
+    for i in 0..zone_count {
+        let u8name_array = [i as u8; CONFIG_NAME_MAXLEN];
+        let zone = Zone::new(i, &u8name_array);
+        ZONE_LIST.write().push(Arc::new(RwLock::new(zone)));
+    }
+    for i in 0..zone_count {
+        remove_zone(i);
+    }
+    assert_eq!(ZONE_LIST.read().len(), zone_count_before);
 }
